Initial import. We can extract a co-authorship network from a table. This extraction does not contain any additional attributes at this time.
We also export the package edu.iu.nwb.analysis.extractmultivaluednetwork, which is the baseline functionality. It allows us to extract networks from arbitrary tables given a column name and a character spliter. Further functionality will be available through the use of a metadata file which will allow us to aggregate data on the columns.

git-svn-id: svn+ssh://in.cns.iu.edu/projects/svn/nwb/trunk@1463 c287de05-5010-0410-818a-8f02fb49d6a9