Added ability to color the inside of circles (and the ability to turn that feature off so that no inner color is drawn).  Changed "circle color" to "circle outer color" to distinguish from this, and added ability to turn that off too (defaults to black).
CirclePrinter now PostScript-draws the circles in order from largest to smallest to minimize the chance of occlusion when the user wants inner colors.

Refactored reading of Table data in RegionAnnotationMode and CircleAnnotationMode.  This should be slightly less ugly.  Removed ListInterpolator and ListScaler in favor of convenience methods in the appropriate classes.

Reviewed by Micah.

git-svn-id: svn+ssh://in.cns.iu.edu/projects/svn/nwb/trunk@3671 c287de05-5010-0410-818a-8f02fb49d6a9