Added the GNU Scientific Library DLLs currently needed for the Windows executable (SPEM.exe) to run.
Replaced SPEM.exe with a compile on source slightly modified from Bruno's.  This is to tide us over until he's back in town and can update his source.  Particular changes, both to SPEM.cpp:
* Two lines around 50 have changed from "stdfoo = bar" to using freopen.  Without this, we get a mingw32-make compile error (non-lvalue in assignment).
* In the input.openFiles around line 78, I have changed the file names from ".foo.dat.gz" to ".foo.dat" so that SPEMShell will not gzip its output.  If necessary we could revert this and gunzip SPEMShell's output in some platform-independent fashion.

Note that these changes are *NOT* reflected in the attached source (SPEM.tgz) since the first is a platform-dependent (for Windows) fix and since the gzip change will hopefully happen on Bruno's end ultimately.

git-svn-id: svn+ssh://in.cns.iu.edu/projects/svn/nwb/trunk@3812 c287de05-5010-0410-818a-8f02fb49d6a9