Added a Python script that reads a calibrated dataset and a (trained, I imagine) codebook and produces a CSV giving the distribution of times each neuron in the codebook was a best-matching unit.
Added in-progress C++ script that reads a calibrated dataset and a (trained, I imagine) codebook and produces a merged codebook.  The idea is to create one "cluster" for each neuron and have each of these clusters "contain" all that documents for which its neuron was the BMU.  Then we iteratively merge pairs of clusters (starting with the "smallest") until they all meet some minimum number of contained documents.

git-svn-id: svn+ssh://in.cns.iu.edu/projects/svn/nwb/trunk@4393 c287de05-5010-0410-818a-8f02fb49d6a9