Switched from Cluster struct to a variety of vectors indexed by integer cluster IDs.
Switched from Coordinate struct to pair<int, int> for recording row and column information.  Added macros ROW() and COLUMN().

Added the heap comparator (for choosing clusters to merge according to how few documents they contain).

Fixed a little bug in readCodebook.  We had been getting an initial empty line.

Corrected an oversight in the combine function.  We had not been marking the newly created/merged cluster as the replacement neighbor for all of the two input clusters' neighbors.

Added detail to output cluster information in the final write to file.

git-svn-id: svn+ssh://in.cns.iu.edu/projects/svn/nwb/trunk@4397 c287de05-5010-0410-818a-8f02fb49d6a9