Added network version of EpiC simulator runner.  It accepts a Model and a file:text/nwb.
Re-organized simulator runner to take advantage of common code between the single-population version and the network version.  Both are stochastic (while the exact version is not), which means they accept as algorithm parameters both a random number generator seed and a number of runs.  These live in edu.iu.epic.simulator.runner.stochastic.

The "population" algorithm parameter is only provided for the exact and single-population versions.  In the network version this figure is determined by the number of nodes in the network.

Added NetworkFileMaker, which adapts an NWB file to the EdgeList-style format expected by the core.

Fixed output filename for edu.iu.epic.simulator.core.network.

More generous tolerance for that the initial population distributions sum to 1.  Rather than checking an exact sum to 1.0f, we allow an error up to 1e-3f.

git-svn-id: svn+ssh://in.cns.iu.edu/projects/svn/nwb/trunk@5059 c287de05-5010-0410-818a-8f02fb49d6a9