On post-processing compartment population time series output from the core simulator executables:
This code previously interpreted those numbers as floats.  This is because the code is used to post-process output from every simulator and while some of them output only integer populations (Single-Population), others (Exact) output fractional ones.

The timestep value was also being rewritten as a float (even though these are always integers), yielding a CSV containing for example 1.0 instead of 1 for the first timestep as so on.

When using the Line Graph visualization algorithm on data with this quirk, the CSV would be converted to a Prefuse table and its columns searched for some column of type Integer.  In the case of the Exact simulator, no such column existed and the Line Graph visualization failed at parameter mutation time.

This rewritten version preserves integers from the core simulator output where possible and uses floats otherwise.

Additionally, and unrelatedly, this code now suppresses any row composed entirely of zeroes (the timestep column is included in the consideration).  I don't know why, but the Exact simulator core produces such a line while the others do not.


Reviewed by Chin Hua.

git-svn-id: svn+ssh://in.cns.iu.edu/projects/svn/nwb/trunk@5387 c287de05-5010-0410-818a-8f02fb49d6a9