Multiple category membership.
The static listing of categories was previously kept in a static initial_data.json within the categories app.  It was missing slugs for each category.  Slugs are generated automatically by the Category model.  So we created a fixture generator initial_data.py in the categories app that will create the listing of categories initial_data.json dynamically from a simple list of category names.  I'm hoping this will keep us in sync better with the Category model.

git-svn-id: svn+ssh://in.cns.iu.edu/projects/svn/nwb/trunk@5453 c287de05-5010-0410-818a-8f02fb49d6a9