In line with the previous commit, this change embeds the fact that the 0002 datasets migration is dependent on the 0002 projects migration in the migration itself rather than by migrating each app individually in an externally imposed order set by migrate_database.sh.
Unfortunately I can't think of a way to test that this is the *only* dependency that requires stating.  The state of the production server has changed so that it is no longer possible to restore the old tag and check this fix in comparison to the original fix.  Further dependency constraints may be required.

git-svn-id: svn+ssh://in.cns.iu.edu/projects/svn/nwb/trunk@5488 c287de05-5010-0410-818a-8f02fb49d6a9