Changed algorithm to conform with NETE specifications (JIRA-3).
The postscript files are now generated by classes that are responsible for parts of the postscript document. This will eventually need to be changed to use templates for improved readability. It would also be good to add a valueclass to reduce the method signatures. The deprecated Date class is used extensively and should be changed to jodatime.

Reviewed by: Joseph 

git-svn-id: svn+ssh://in.cns.iu.edu/projects/svn/nwb/trunk@5604 c287de05-5010-0410-818a-8f02fb49d6a9