Preparation for http://jira.cns.iu.edu/browse/SCISQUARED-1011:
DisciplineBreakdownAreas.Page initialized a List field with size equal to an uninitialized int field.  In effect this List field had initial size zero.  Replaced this initialization with a clearer equivalent.

TableReader#TableReader(Table, String, String, String, LogService) contained an apparent bug near line 110.  A Map<Integer, Float> was checking #containsKey on a String argument.  Presumably the containsKey check was meant to be run on the same Map on which #get is called in the conditional body, which is of type Map<String, Float>.

CircleSizeLegend#render(GraphicsState) contained an insignificant integer division bug in display position calculation.

Sending exceptions caught but not rethrown to LOG_DEBUG.

Missing @Override annotations.

Reviewed by Thomas.

git-svn-id: svn+ssh://in.cns.iu.edu/projects/svn/nwb/trunk@6328 c287de05-5010-0410-818a-8f02fb49d6a9